In search for ultimate accuracy
So chunky and cumulative commit finishing quite long thoughts,
experiments and learnings of improving crosh algorithm, so that
it may perhaps have ultimate accuracy even in middle cases of
rounding because of some complex math (a.k.a. magic). Now it's
clear in theory that it's impossible, which is quite expected
looking at grisu limitations in almost the same border cases.

It's also clear now that all the dword_* logic, guessing code,
all the tries to increase precision by a little bit should
be dropped.

According to tests, even precise diy_fp multiplication should
be removed in favour of coarse one used in grisu.

Most of excessive code found in the previous commits was
removed/commented out, except for precision and dword code
kept here just as an example.

In general, commit includes the following changes:
1. Code clean-up and several fixes with bit operations.
2. Bring atof() function to conformity with strtod()-family API-wise.
3. Introduce fallback conversion which is called in border cases
with unclear rounding mentioned above.
4. Improve round_up() test.
5. Keep dword_* logic and precise diy_fp multiplication for now.
6. Remove/comment all the guessing and surrogate-precision-increase-
by-a-couple-of-bits code.